♻️ Séparation des edit modes dans plusieurs modules
