+++ src/app/mod.rs
@@ -1,12 +1,14 @@
 use crate::ui::ui;
-use commands::{check_cmd, CommandLine};
+use commands::CommandLine;
+use editor_modes::{command_mode, insert_mode, normal_mode};
 use ratatui::{
     backend::Backend,
-    crossterm::event::{self, Event, KeyCode, KeyEvent},
+    crossterm::event::{self, Event, KeyEvent},
     Terminal,
 };
 use std::io::Result;
 pub mod commands;
+mod editor_modes;
 
 #[derive(PartialEq, Eq)]
 pub enum InputMode {
@@ -28,49 +30,26 @@ impl App {
             command_line: CommandLine::new(),
         }
     }
-    fn manage_input_mode(self: &mut App, key: KeyEvent) {
+
+    fn handle_modes(self: &mut App, key: KeyEvent) {
         match self.input_mode {
-            InputMode::Normal => match key.code {
-                KeyCode::Char('i') => self.input_mode = InputMode::Insert,
-                KeyCode::Char(':') => {
-                    self.input_mode = InputMode::Command;
-                    self.command_line.command_buffer.push(':');
-                }
-                _ => {}
-            },
-            InputMode::Insert => match key.code {
-                KeyCode::Esc => self.input_mode = InputMode::Normal,
-                _ => {}
-            },
-            InputMode::Command => match key.code {
-                KeyCode::Char(insert) => {
-                    self.command_line.command_buffer.push(insert);
-                }
-                KeyCode::Enter => {
-                    self.input_mode = InputMode::Normal;
-                    check_cmd(self);
-                    self.command_line.command_buffer = "".to_string();
-                }
-                KeyCode::Esc => {
-                    self.input_mode = InputMode::Normal;
-                    self.command_line.command_buffer = "".to_string();
-                }
-                _ => {}
-            },
+            InputMode::Normal => normal_mode(self, key.code),
+            InputMode::Insert => insert_mode(self, key.code),
+            InputMode::Command => command_mode(self, key.code),
             _ => {}
         }
     }
-}
 
-pub fn run_app<B: Backend>(terminal: &mut Terminal<B>, app: &mut App) -> Result<()> {
-    loop {
-        terminal.draw(|frame| ui(frame, app))?;
-        if let Event::Key(key) = event::read()? {
-            app.manage_input_mode(key);
+    pub fn run_app<B: Backend>(self: &mut App, terminal: &mut Terminal<B>) -> Result<()> {
+        loop {
+            terminal.draw(|frame| ui(frame, self))?;
+            if let Event::Key(key) = event::read()? {
+                self.handle_modes(key);
+            }
+            if self.input_mode == InputMode::Exit {
+                break;
+            }
         }
-        if app.input_mode == InputMode::Exit {
-            break;
-        }
+        Ok(())
     }
-    Ok(())
 }
